requirements.txt
fastapi
python-dotenv
langchain_experimental
langchain_core
langchain-google-genai
setuptools
pydantic
langchain_chroma